--keep-git

+++ bin/get-template.js
@@ -5,14 +5,18 @@ const axios = require('axios');
 
 const ALIAS_URL = 'https://raw.githubusercontent.com/get-template/aliases/main/aliases.txt';
 
+const FLAGS = ['--ci', '--execute-commands', '-ec']
+
 async function getTemplate() {
 	const args = process.argv.slice(2);
 
 	const template = args[0];
-	const directory = (!args[1] || ['--ci', '--execute-commands', '-ec'].includes(args[1]) ? '.' : args[1]);
+	const directory = (!args[1] || FLAGS.includes(args[1]) ? '.' : args[1]);
 
+	// Flags
 	const ci = !!args.find(flag => flag === '--ci');
 	const executeCommands = !!args.find(flag => flag === '--execute-commands' || flag === '-ec');
+	const removeGit = !args.find(flag => flag === '--keep-git');
 
 	let gitDestination;
 	let postDownloadCommand;
@@ -92,25 +96,28 @@ async function getTemplate() {
 	// Download template
 	try {
 		console.log(`Downloading template from ${gitDestination}...`);
-		await exec(`git clone ${gitDestination} ${directory}`);
+		await exec(`git clone ${gitDestination} ${directory} ${removeGit && '--depth 1'}`);
 	} catch (e) {
 		return error('Could not clone repository', e);
 	}
 	
-	// Remove .git
-	try {
-		console.log('Removing .git...');
-		await exec(`cd ${directory}; rm -rf .git`);
-	} catch (e) {
-		return error('Could not remove .git', e);
-	}
+	// Git operations
+	if (removeGit) {
+		// Remove .git
+		try {
+			console.log('Removing .git...');
+			await exec(`cd ${directory}; rm -rf .git`);
+		} catch (e) {
+			return error('Could not remove .git', e);
+		}
 
-	// Initialize .git
-	try {
-		console.log('Initializing new git...');
-		await exec(`cd ${directory}; git init`);
-	} catch (e) {
-		return error('Could not initialize git', e);
+		// Initialize .git
+		try {
+			console.log('Initializing new git...');
+			await exec(`cd ${directory}; git init`);
+		} catch (e) {
+			return error('Could not initialize git', e);
+		}
 	}
 
 	// Execute command
